Refactor: Refactored A Bunch Of Classes

+++ voxel-engine-frustum-culling/src/main/java/com/duckyshine/app/physics/ray/RayResult.java
@@ -1,28 +1,28 @@
 package com.duckyshine.app.physics.ray;
 
-import org.joml.Vector3i;
+import org.joml.Vector3f;
 
 public class RayResult {
     private boolean isIntersect;
 
-    private Vector3i axes;
-    private Vector3i position;
+    private Vector3f axes;
+    private Vector3f position;
 
     public RayResult() {
-        this.axes = new Vector3i();
+        this.axes = new Vector3f();
         this.position = null;
     }
 
-    public RayResult(Vector3i position, Vector3i axes) {
+    public RayResult(Vector3f position, Vector3f axes) {
         this.axes = axes;
         this.position = position;
     }
 
-    public void setAxes(Vector3i axes) {
+    public void setAxes(Vector3f axes) {
         this.axes = axes;
     }
 
-    public void setPosition(Vector3i position) {
+    public void setPosition(Vector3f position) {
         this.position = position;
     }
 
@@ -34,11 +34,11 @@ public class RayResult {
         return this.isIntersect;
     }
 
-    public Vector3i getAxes() {
+    public Vector3f getAxes() {
         return this.axes;
     }
 
-    public Vector3i getPosition() {
+    public Vector3f getPosition() {
         return this.position;
     }
 }
